api/requirements-dev.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
wikipedia==1.4.0
redis==5.2.1
APScheduler==3.11.0
instagrapi==2.1.3
cryptography==44.0.0
pillow==11.1.0
markdown-it-py==3.0.0
mdit-plain==1.0.1
aiohttp==3.11.14
